+++ .eslintrc.json
@@ -27,7 +27,8 @@
       "react-native/no-inline-styles": 0,
       "no-shadow": 0,
       "react/react-in-jsx-scope": 0,
-      "react/no-unescaped-entities": 0
+      "react/no-unescaped-entities": 0,
+      "quotes": [2, "double", { "avoidEscape": true }]
     },
     "ignorePatterns": ["src/library/*"]
   }
