+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-  "singleQuote": true,
+  "singleQuote": false,
   "trailingComma": "none",
   "tabWidth": 2,
   "semi": false,
