+++ pages/index.tsx
@@ -1,24 +1,11 @@
-import { Box, Container, Grid, Typography } from '@mui/material'
-import Head from 'next/head'
-import Image from 'next/image'
-import Ticket from '../components/UI/ticket/Ticket'
-import { motion, useAnimation } from 'framer-motion'
-import { grey } from '@mui/material/colors'
-import ContainedButton from '../components/UI/button/ContainedButton'
-import AnimateWhenInView from '../components/motion/AnimateWhenInView'
-import { useRouter } from 'next/navigation'
-import { height } from '@mui/system'
-import ContainerCard from '../components/UI/card/ContainerCard'
-
-const background1 =
-  'radial-gradient(116.3% 117.1% at 93.9% 40.4%, rgb(43, 46, 49) 0%, rgb(0, 0, 0) 100%);'
-const background2 =
-  'radial-gradient(116.3% 117.1% at 93.9% 40.4%, rgb(146, 196, 255) 0%, rgb(255, 255, 255) 100%);'
+import { Box, Container } from "@mui/material"
+import Head from "next/head"
+import LandingTopSection from "../components/landing/LandingTopSection"
+import IssuesSection from "../components/landing/IssuesSection"
+import ObjectivesSection from "../components/landing/ObjectivesSection"
+import GetStartedSection from "../components/landing/GetStartedSection"
 
 export default function Home() {
-  const controls = useAnimation()
-  const router = useRouter()
-
   return (
     <Container>
       <Head>
@@ -26,379 +13,20 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <Box
-        sx={{
-          left: '50%',
-          right: '50%',
-          marginLeft: '-50vw',
-          marginRight: '-50vw',
-          display: 'flex',
-          justifyContent: 'center',
-          overflow: 'hidden'
-        }}
-      >
-        <Image
-          src="/images/landing/background.webp"
-          height="600"
-          width="2500"
-          alt="background"
-          style={{
-            position: 'absolute',
-            left: 0,
-            objectFit: 'cover',
-            width: '100vw',
-            maxWidth: '100vw'
-          }}
-          draggable={false}
-        />
-        <Grid
-          container
-          justifyContent="center"
-          sx={{
-            position: 'absolute',
-            width: '100vw',
-            left: 0,
-            right: 0,
-            margin: '0 auto'
-          }}
-        >
-          <Grid xs={5} lg={5} maxWidth="xl" sx={{}}>
-            <Box sx={{ height: 120 }} />
-            <Typography variant="h3" component="h1">
-              Warden
-            </Typography>
-            <Box sx={{ height: 12 }} />
-            <Typography component="p" color={grey[700]}>
-              A one-stop event-ticketing platform that mitigates the issue of
-              ticket authenticity, and the lack of secondary market control over
-              ticket transactions
-            </Typography>
-            <Box sx={{ height: 20 }} />
-            <ContainedButton
-              onClick={() => {
-                router.push('/home')
-              }}
-              disabled={false}
-              variant="contained"
-              label="Get started"
-              height="40px"
-              width="200px"
-            />
-          </Grid>
-          <Grid
-            xs={5}
-            lg={5}
-            maxWidth="xl"
-            sx={{
-              position: 'relative',
-              display: 'flex',
-              justifyContent: 'center'
-            }}
-          >
-            <motion.div
-              animate={controls}
-              transition={{ staggerChildren: 0.5 }}
-            >
-              <motion.div
-                key={1}
-                initial={{ y: 100, opacity: 0 }}
-                animate={{ y: 0, opacity: 0.6 }}
-                transition={{
-                  delay: 0.7,
-                  duration: 0.6,
-                  type: 'spring',
-                  bounce: 0.4
-                }}
-                id="card1"
-              >
-                <Ticket
-                  img="/images/background/ticket-cover-test.jpg"
-                  eventName="Among Us Party"
-                  eventOrganizer="Warden"
-                  ticketType="GENERAL"
-                  date={new Date()}
-                  seat="A 14"
-                  location="MIRA HQ"
-                  QRCodeValue="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
-                />
-              </motion.div>
-              <motion.div
-                key={2}
-                initial={{ y: 100, opacity: 0 }}
-                animate={{ y: 0, opacity: 0.8 }}
-                transition={{
-                  delay: 0.6,
-                  duration: 0.5,
-                  type: 'spring',
-                  bounce: 0.4
-                }}
-                id="card2"
-              >
-                <Ticket
-                  img="/images/background/ticket-cover-test.jpg"
-                  eventName="Among Us Party"
-                  eventOrganizer="Warden"
-                  ticketType="GENERAL"
-                  date={new Date()}
-                  seat="A 14"
-                  location="MIRA HQ"
-                  QRCodeValue="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
-                />
-              </motion.div>
-              <motion.div
-                key={3}
-                initial={{ y: 100, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{
-                  delay: 0.5,
-                  duration: 0.3,
-                  type: 'spring',
-                  bounce: 0.4
-                }}
-                id="card3"
-              >
-                <Ticket
-                  img="/images/background/ticket-cover-test.jpg"
-                  eventName="Among Us Party"
-                  eventOrganizer="Warden"
-                  ticketType="GENERAL"
-                  date={new Date()}
-                  seat="A 14"
-                  location="MIRA HQ"
-                  QRCodeValue="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
-                />
-              </motion.div>
-            </motion.div>
-          </Grid>
-        </Grid>
-      </Box>
-      <Box sx={{ height: '600px' }} />
+      {/* Landing Top Section */}
+      <LandingTopSection />
       <Box sx={{ height: 144 }} />
-      <Box sx={{ maxWidth: '1200px' }}>
-        <Typography variant="h4" component="h4">
-          Issues with the ticketing industry
-        </Typography>
-        <Box sx={{ height: 24 }} />
-        <Box
-          sx={{
-            width: '100%',
-            display: 'flex',
-            justifyContent: 'center',
-            gap: 4
-          }}
-        >
-          <Box
-            sx={{
-              background: background1,
-              borderRadius: 4,
-              height: 300,
-              width: '50%',
-              padding: 6,
-              color: '#fff'
-            }}
-          >
-            <AnimateWhenInView>
-              <Image
-                src="/images/landing/trust.png"
-                width={100}
-                height={100}
-                alt="trust"
-              />
-            </AnimateWhenInView>
-            <Box sx={{ height: 24 }} />
-            <Typography variant="h6" component="h6">
-              Issues of Authenticity and Customer Trust
-            </Typography>
-            <Box sx={{ height: 24 }} />
-            <Typography component="p">
-              Need to track the authenticity and history of tickets
-            </Typography>
-            <Typography component="p">
-              Event organizers need to verify the authenticity of the ticket and
-              ticket holders
-            </Typography>
-          </Box>
-          <Box
-            sx={{
-              background: background2,
-              borderRadius: 4,
-              height: 300,
-              width: '50%',
-              padding: 6,
-              color: '#000'
-            }}
-          >
-            <AnimateWhenInView>
-              <Image
-                src="/images/landing/control.png"
-                width={100}
-                height={100}
-                alt="trust"
-              />
-            </AnimateWhenInView>
-            <Box sx={{ height: 24 }} />
-            <Typography variant="h6" component="h6">
-              Issues of Authenticity and Customer Trust
-            </Typography>
-            <Box sx={{ height: 24 }} />
-            <Typography component="p">
-              Need to track the authenticity and history of tickets
-            </Typography>
-            <Typography component="p">
-              Event organizers need to verify the authenticity of the ticket and
-              ticket holders
-            </Typography>
-          </Box>
-        </Box>
-      </Box>
 
-      <Box sx={{ height: 120 }} />
-      <Box sx={{ maxWidth: '1200px' }}>
-        <Typography variant="h4" component="h4">
-          Our Objectives
-        </Typography>
-        <Box sx={{ height: 24 }} />
-        <Box sx={{ display: 'flex', gap: 2, justifyContent: 'space-between' }}>
-          <ContainerCard>
-            <>
-              <Typography variant="h6" fontWeight="bold">
-                Authenticity of tickets
-              </Typography>
-              <Typography>Authenticity of tickets</Typography>
-              <Typography>Transparency in the system</Typography>
-            </>
-          </ContainerCard>
-          <ContainerCard>
-            <>
-              <Typography variant="h6" fontWeight="bold">
-                Bot Reduction
-              </Typography>
-              <Typography>
-                Limit tickets per person with Smart Contract
-              </Typography>
-              <Typography>Identity Binding</Typography>
-            </>
-          </ContainerCard>
-          <ContainerCard>
-            <>
-              <Typography variant="h6" fontWeight="bold">
-                Marketplace Control
-              </Typography>
-              <Typography>Native Secondary Market</Typography>
-              <Typography>
-                Transactions regulated with Smart Contract
-              </Typography>
-            </>
-          </ContainerCard>
-        </Box>
-        <Box sx={{ height: 20 }} />
-        <Box sx={{ display: 'flex', gap: 2, justifyContent: 'space-between' }}>
-          <ContainerCard>
-            <>
-              <Typography variant="h6" fontWeight="bold">
-                Redistribute Income
-              </Typography>
-              <Typography>
-                Event Organizers earn the royalty fee from ticket resales
-              </Typography>
-            </>
-          </ContainerCard>
-          <ContainerCard>
-            <>
-              <Typography variant="h6" fontWeight="bold">
-                Utilizing Tickets
-              </Typography>
-              <Typography>
-                View tickets and their transaction history
-              </Typography>
-              <Typography>Tickets for event admisison</Typography>
-            </>
-          </ContainerCard>
-          <ContainerCard>
-            <>
-              <Typography variant="h6" fontWeight="bold">
-                User-friendliness
-              </Typography>
-              <Typography>
-                User-friendly for beginner and expert users
-              </Typography>
-            </>
-          </ContainerCard>
-        </Box>
-      </Box>
-
-      {/* NEW SECTION */}
+      {/* Issues Section */}
+      <IssuesSection />
       <Box sx={{ height: 144 }} />
-      <Box
-        sx={{
-          left: '50%',
-          right: '50%',
-          marginLeft: '-50vw',
-          marginRight: '-50vw',
-          display: 'flex',
-          justifyContent: 'center',
-          overflow: 'hidden'
-        }}
-      >
-        <Box sx={{ height: 220 }} />
-        <Box
-          style={{
-            position: 'absolute',
-            left: 0,
-            right: 0,
-            objectFit: 'cover',
-            height: '240px',
-            background: '#1f1f1f'
-          }}
-        />
-        <Box
-          sx={{
-            position: 'absolute',
-            width: '100vw',
-            display: 'flex',
-            justifyContent: 'center'
-          }}
-        >
-          <Box
-            sx={{
-              position: 'absolute',
-              top: -20,
-              right: 0,
-              left: 0,
-              display: 'flex',
-              justifyContent: 'center'
-            }}
-          >
-            <ContainedButton
-              onClick={() => {
-                router.push('/home')
-              }}
-              disabled={false}
-              variant="contained"
-              label="Get started"
-              height="40px"
-              width="200px"
-            />
-          </Box>
-          <Box
-            sx={{
-              marginTop: 8,
-              width: '600px',
-              display: 'flex',
-              justifyContent: 'center'
-            }}
-          >
-            <Typography color="#fff" align="center" fontSize={16}>
-              Warden aims to develop a unified platform of an NFT collection
-              generator, a native marketplace, and an events management system
-              to address these issues using the capabilities of NFT and
-              blockchain technology. We aim to bring back trust to the
-              customers, regain control of the secondary market, and redirect
-              ticket resale revenues back to the event organizers.
-            </Typography>
-          </Box>
-        </Box>
-      </Box>
+
+      {/* Objectives Section */}
+      <ObjectivesSection />
+      <Box sx={{ height: 144 }} />
+
+      {/* Get Started Section */}
+      <GetStartedSection />
     </Container>
   )
 }
